BTEAM-141 [Update] エネルギー小の場合のダメージ追加

+++ DateApps2023/Assets/Project/Scripts/Boss/BossMove.cs
@@ -9,7 +9,7 @@ public class BossMove : MonoBehaviour
 
 
     [SerializeField]
-    public int bossHp;
+    public float bossHp;
 
     private BossAttack bossAttack;
     private BossDamage bossDamage;
